refactor: small adjustments.
