+++ Assets/Scripts/AI/AIInputController.cs
@@ -1,5 +1,6 @@
-﻿using LeandroExhumed.SnakeGame.Grid;
-using System;
+﻿using LeandroExhumed.SnakeGame.Block;
+using LeandroExhumed.SnakeGame.Grid;
+using LeandroExhumed.SnakeGame.Match;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -11,19 +12,22 @@ namespace LeandroExhumed.SnakeGame.AI
         private readonly AIInputView view;
 
         private readonly IGridModel<INodeModel> levelGrid;
+        private readonly IMatchModel match;
 
-        public AIInputController (IAIInputModel model, AIInputView view, IGridModel<INodeModel> levelGrid)
+        public AIInputController (IAIInputModel model, AIInputView view, IGridModel<INodeModel> levelGrid, IMatchModel match)
         {
             this.model = model;
             this.view = view;
             this.levelGrid = levelGrid;
+            this.match = match;
         }
 
         public void Setup ()
         {
             model.OnPathChanged += HandlePathChanged;
-            levelGrid.OnNodeChanged += HandleLevelGridNodeChanged;
             model.OnDestroyed += HandleDestroyed;
+            levelGrid.OnNodeChanged += HandleLevelGridNodeChanged;
+            match.OnBlockGenerated += HandleBlockGenerated;
         }
 
         private void HandlePathChanged (List<IPathNodeModel> path)
@@ -41,11 +45,17 @@ namespace LeandroExhumed.SnakeGame.AI
             model.HandleGridNodeChanged(nodePosition);
         }
 
+        private void HandleBlockGenerated (IBlockModel block)
+        {
+            model.HandleBlockGenerated(block);
+        }
+
         public void Dispose ()
         {
             model.OnPathChanged -= HandlePathChanged;
-            levelGrid.OnNodeChanged -= HandleLevelGridNodeChanged;
             model.OnDestroyed -= HandleDestroyed;
+            levelGrid.OnNodeChanged -= HandleLevelGridNodeChanged;
+            match.OnBlockGenerated -= HandleBlockGenerated;
         }
     }
 }
