+++ Assets/Scripts/AI/AIInputModel.cs
@@ -38,14 +38,11 @@ namespace LeandroExhumed.SnakeGame.AI
         private readonly IPathFindingModel pathFinding;
         private readonly MonoBehaviour monoBehaviour;
 
-        private readonly IMatchModel match;
-
         public AIInputModel (AIData data, IPathFindingModel pathFinding, MonoBehaviour monoBehaviour, IMatchModel match)
         {
             this.data = data;            
             this.pathFinding = pathFinding;
             this.monoBehaviour = monoBehaviour;
-            this.match = match;
         }
 
         public void Initialize (ISnakeModel snake)
@@ -57,7 +54,7 @@ namespace LeandroExhumed.SnakeGame.AI
         public void Initialize ()
         {
             snake.OnPositionChanged += HandlePositionChanged;
-            match.OnBlockGenerated += HandleBlockGenerated;
+            
         }
 
         public void HandleGridNodeChanged (Vector2Int nodePosition)
@@ -76,6 +73,25 @@ namespace LeandroExhumed.SnakeGame.AI
                     data.MaxReasoningTimeToPlanPathAfterObstruction);
                 monoBehaviour.StartCoroutine(ThinkAboutNewPath(reasoningTime));
             }
+        }
+
+        public void HandleBlockGenerated (IBlockModel block)
+        {
+            if (targetBlock != null)
+            {
+                float currentDistance = Vector2Int.Distance(snake.Position, targetBlock.Position);
+                if (currentDistance < Vector2Int.Distance(snake.Position, block.Position))
+                {
+                    return;
+                }
+            }
+
+            targetBlock = block;
+            reasonedAboutNewBlock = false;
+            float reasoningTime = UnityEngine.Random.Range(
+                data.MinReasoningTimeToPlanPathToBlock,
+                data.MaxReasoningTimeToPlanPathToBlock);
+            monoBehaviour.StartCoroutine(ThinkAboutNewPath(reasoningTime));
         }
 
         public void HandleBlockCollected (IBlockModel block)
@@ -159,29 +175,10 @@ namespace LeandroExhumed.SnakeGame.AI
             monoBehaviour.StartCoroutine(RequestMovementInput(input));
         }
 
-        private void HandleBlockGenerated (IBlockModel block)
-        {
-            if (targetBlock != null)
-            {
-                float currentDistance = Vector2Int.Distance(snake.Position, targetBlock.Position);
-                if (currentDistance < Vector2Int.Distance(snake.Position, block.Position))
-                {
-                    return;
-                }
-            }
-
-            targetBlock = block;
-            reasonedAboutNewBlock = false;
-            float reasoningTime = UnityEngine.Random.Range(
-                data.MinReasoningTimeToPlanPathToBlock,
-                data.MaxReasoningTimeToPlanPathToBlock);
-            monoBehaviour.StartCoroutine(ThinkAboutNewPath(reasoningTime));
-        }
-
         public void Dispose ()
         {
             snake.OnPositionChanged -= HandlePositionChanged;
-            match.OnBlockGenerated -= HandleBlockGenerated;
+            
         }
     }
 }
